requirements.txt
django~=4.1.3
djangorestframework~=3.14.0
psycopg2-binary
django-filter~=22.1

